Fix issues with db and web services. Add debug version of the web request. Add detailsView and finish MVP

+++ SpoonacularApp/Network/DTOs/DetailedRecipe/MetricAPI.swift
@@ -3,7 +3,7 @@ import Foundation
 
 struct MetricAPI: Codable {
 
-  var amount    : Int?    = nil
+  var amount    : Double?    = nil
   var unitShort : String? = nil
   var unitLong  : String? = nil
 
@@ -18,7 +18,7 @@ struct MetricAPI: Codable {
   init(from decoder: Decoder) throws {
     let values = try decoder.container(keyedBy: CodingKeys.self)
 
-    amount    = try values.decodeIfPresent(Int.self    , forKey: .amount    )
+    amount    = try values.decodeIfPresent(Double.self    , forKey: .amount    )
     unitShort = try values.decodeIfPresent(String.self , forKey: .unitShort )
     unitLong  = try values.decodeIfPresent(String.self , forKey: .unitLong  )
