+++ SpoonacularApp/Network/DTOs/DetailedRecipe/PropertiesAPI.swift
@@ -4,7 +4,7 @@ import Foundation
 struct PropertiesAPI: Codable {
 
   var name   : String? = nil
-  var amount : Int?    = nil
+  var amount : Double?    = nil
   var unit   : String? = nil
 
   enum CodingKeys: String, CodingKey {
@@ -19,7 +19,7 @@ struct PropertiesAPI: Codable {
     let values = try decoder.container(keyedBy: CodingKeys.self)
 
     name   = try values.decodeIfPresent(String.self , forKey: .name   )
-    amount = try values.decodeIfPresent(Int.self    , forKey: .amount )
+    amount = try values.decodeIfPresent(Double.self    , forKey: .amount )
     unit   = try values.decodeIfPresent(String.self , forKey: .unit   )
  
   }
